Optionally enabling dashboard page autorefresh.
If a $settings.dashboard.autorefresh value is specified, a
http-equiv="refresh" meta tag will be rendered, encouraging
modern browsers to refresh at the specified interval.

There is no validation of this setting, which typically
should be the refresh interval in seconds. An invalid
value should not cause any issues, other than not enabling
browser auto-refresh.