Rich event output web UI
This patch attempts to parse check output as JSON, and, if applicable
render rich event detail to the web UI for events.

Specifically, if the check output is JSON and contains a 'url' attribute,
the url attribute will be rendered as a hyperlink in the event details
modal.  This is useful for checks that inherently have an associated URL,
i.e. an HTTP request check, or RPC failure that has an web UI, ala resque.

Additionally, this patch will attempt to render the 'message' attribute
of the output, if present, and fallback to rending the raw output. This
is necesary because raw json is hard to parse visually in the web UI.